style: rTK
Conversion of the traditional Redux to RTK pattern
