+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.3.6",
     "@testing-library/jest-dom": "5.9.0",
     "@testing-library/react": "10.0.6",
     "@testing-library/user-event": "11.0.0",
@@ -22,9 +23,9 @@
     "redux": "4.0.5",
     "redux-persist": "6.0.0",
     "redux-thunk": "2.3.0",
+    "snyk": "1.334.0",
     "typescript": "3.9.3",
-    "uuid": "8.1.0",
-    "snyk": "1.334.0"
+    "uuid": "8.1.0"
   },
   "devDependencies": {
     "@types/react-redux": "7.1.9",
