+++ src/components/AddTodoForm/index.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { Form, Row, Col, Button, Input } from 'antd';
 import { PlusCircleFilled } from '@ant-design/icons';
-
 import './styles.less';
-import { Todo } from 'store/todo/models/todo.model';
+import { Todo } from 'store/todo';
 
 interface AddTodoFormProps {
   onFormSubmit: (todo: Todo) => void;
