+++ src/store/app.store.ts
@@ -1,16 +1,23 @@
-import { createStore, applyMiddleware, Store } from 'redux';
+import { createStore, applyMiddleware, Store, combineReducers } from 'redux';
 import thunk from 'redux-thunk';
 import { persistStore, persistReducer, PersistConfig } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
 import autoMergeLevel2 from 'redux-persist/lib/stateReconciler/autoMergeLevel2';
-
-import { rootReducer } from 'store/todo/reducers';
+import { todoSlice } from 'store/todo';
 
 const reduxPersistConfig: PersistConfig<any> = {
   key: 'application',
   storage: storage,
   stateReconciler: autoMergeLevel2
 };
+
+
+
+export const rootReducer = combineReducers({
+  todo: todoSlice.reducer,
+})
+
+export type RootState = ReturnType<typeof rootReducer>;
 
 const pReducer = persistReducer(reduxPersistConfig, rootReducer);
