--- src/store/todo/constants/index.ts
@@ -1,1 +1,0 @@
-export * from './todo.constants';
