--- src/store/todo/constants/todo.constants.ts
@@ -1,5 +1,0 @@
-export const TODO_CONSTANTS = {
-  ADD_TODO_SUCCESS: 'ADD_TODO_SUCCESS',
-  TOGGLE_TODO_SUCCESS: 'TOGGLE_TODO_SUCCESS',
-  REMOVE_TODO_SUCCESS: 'REMOVE_TODO_SUCCESS'
-};
