--- src/store/todo/models/todo.model.ts
@@ -1,5 +1,0 @@
-export interface Todo {
-  id?: string;
-  name: string;
-  completed?: boolean;
-}
